examples/YOLOv8_pothole_detector_example/requirements.txt
# Base Packages
gunicorn>=20.0.0
uvicorn>=0.20.0
fastapi>=0.87.0
python-multipart>=0.0.6


#############################################
# Add your own dependencies here (optional) #
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
#tqdm>=4.64.0
#pydantic>=1.10.2
ultralytics[YOLO]
omegaconf>=2.3.0
#Flask>=2.0.1
#tensorboard>=2.4.1
